+++ Vault-Coder/src/utopia/coder/vault/model/data/ClassReferences.scala
@@ -8,7 +8,7 @@ import utopia.flow.view.template.Extender
   * @author Mikko Hilpinen
   * @since 14.10.2021, v1.2
   */
-case class ClassReferences(model: ClassModelReferences, dbFactory: Reference, dbModel: Reference,
+case class ClassReferences(targetClass: Class, model: ClassModelReferences, dbFactory: Reference, dbModel: Reference,
                            genericUniqueAccessTrait: Option[Reference], genericManyAccessTrait: Option[Reference],
                            generic: Option[GenericClassReferences])
 	extends Extender[ClassModelReferences]